requirements.txt
streamlit>=1.32
plotly>=5.18
requests>=2.31
pandas>=2.0
python-dotenv>=1.0
python-dateutil>=2.8
